Vector changes (#18)
* Changed generation system to game_board.py

Changed the system of how maps are generated.  A docstring providing how it works will be added later

* Updated __help method

Renamed __help to __help_populate() and added more functionality to it for checking for occupied_by.

* Updated game_board.py

This isn't done yet, but changes were made to include Vector in some areas.

* Updated game_board

Added detailed documentation to explain the new system that was implemented for generation. Also added a method to go through the occupied_by chain in avatars() and stations().

* Updated game_board

Added properties for the field variables.

* Finalized game_board

Consolidated avatars() and stations() into   get_objects(). Made generate map() a stand-alone method as well and use it in generate_game.py.

* Added vector changes

Created new methods in the Vector class that help add (x, y) values to existing vectors. Added these new functions to the files that use coordinates/positions.