Item inventory features (#21)
* Inventory changes

Added documentation and implementation for the inventory for the avatar class. Added appropriate enums and item implementation.

* Inventory controller added

Added an inventory controller that allows the player to select an item from their inventory and set that to be the held item

* Resolved conversation notes

Made the changes wanted.

* Type hinted

I fixed more type hinting

* Resolved discussions

Fixed issues in the inventory_controller to raise errors for bad input.