Avatar improvements (#36)
* Made adjustments

Made adjustments for the tests

* Fixing Avatar

Made some changes with avatar's inventory.

* Accommodations for new inventory

Made changes to the test files to accommodate for the new inventory in the avatar class.

* Avatar implementations

Added what should be final avatar implementations. Last thing is cleaning up documentation to line up with the changes.

* Removed comment

Removed unnecessary comment that was confusing.