Avatar changes (#7)
* Set up

Doing basic set ups to start implementation.

* Tests created

Made tests for some of the new parts of the Avatar class.

* New tests

These are the minor document and test changes from revamp to the avatar class and its testing files.

* Committing to merge

Committing changes to merge master into this branch.

* added steal rate to tech tree

* Revert "added steal rate to tech tree"

This reverts commit 04b27e858118726f980f765071d74518e5524ff1.

* More functionality

Added functionality for the unlock functions for the tech_tree and created tests.

* Tech Tree implementations

Added more tech tree functionality. Some thing might have to be changed as the game's development continues, but it provides a decent foundation.

* PR Changes

Start making the changes requested in the PR.

* PR Changes

Completed requested changes.

* Fixed unused code

Removed unused code.

---------

Co-authored-by: brandongasser <bgasser12@gmail.com>