Game board dynamite list (#18)
* Dynamite List

Started changes to incorporate the dynamite list into the game_board.

* working on test for dynamite_active_ability

* wip

* Dynamite List

Started the process for making the dynamite list. It's almost done in this stage and needs the test file to be done.

* wip

* moving these tests elsewhere

* Dynamite List

Dynamite List should be done. There are some other changes in other files to help with dynamite itself. **I might be missing something in the json test.** It doesn't check if the objects equal, but the size of the lists are correct.

* Fixed JSON

Fixed the JSON tests!!

* Fixed Files

Fixed files that had issues with inventory manager. Scope should be better for trap and dynamite list

---------

Co-authored-by: MyFridgeIsFinePleaseStopAsking <101206316+JuliaCaesar@users.noreply.github.com>