Add dynamite exploding (#38)
* Starting progress

Started implementing the exploding dynamite

* Occupiable tests

Added new test to the occupiable class that will remove a GameObject from the stack given a GameObject.

* Testing

Tested the dynamite controller.

* Controller connects to gameboard

The DynamiteController is placed in the GameBoardController class in the pre_tick() method. This avoid circular imports and helps separate the explosions from the clean up of the dynamite list

* Place controller updates (#39)

* Initial changes

Made the initial changes for preventing multiple dynamite or trap objects from being on a tile

* Tested implementations

Tested the new implementations, and the player cannot place duplicate traps and dynamite on a tile

* New test

Made a new test that ensures dynamite and traps can be on the same tile, but no duplicates of either are allowed.