Competiton docs (#56)
* Initial setup

Changed the previous doc folder name to be 'engine_docs' to separate it from the game documentation. Initial set up with new docs folder started.

* Beginning of website

Made the first few pages for the website

* Gameboard and Ores

Added pages for Gameboard and Ores

* Placeables page

Added a page to explain the placeable objects. Included gifs

* Mobbot page and updates

Made the mobbot page and updated some of the other ones to look nicer

* Fixed mobbot page

Fixed mobbot page to be more user friendly

* Mobbot page updates

Added images to the mobbot page and made it look nicer

* Scoring page

Started work on a page that will show all the scoring in one place

* Theme update

Updated the theme to allow for dark mode, and made small changes

* Tech name change

Changed the tech names to match what will officially be the new tech names

* Test fix

Fixed a test in the dynamite test class after merging master.

* Updates

Updated the ores and scoring pages

* Enums page

Added a page to display the different enums that will be useful to the competitors. May need to expand

* Abilities explained

Explained the difference between the active and passive abilities

* Visualizer page

Added a page to explain the visualizer

* Update mobbot.rst

Explained the inventory system in the mobbot page

* Document clean up

Cleaned up the documents slightly and removed personal notes on what to do

* Start new page

Started the new taking action page

* Update taking_action.rst

Added content to the taking actions page

* Server page

Added a new server page to explain server commands

* Recommended changes

Made some changes to improve QOL

* Fixed gitignore

Removed the build file from the gitignore for docs

* Changes

Made changes in taking action to explain how buying techs work

* Game balance

Added the values for game balancing. Was missed earlier

* Requested changes

Made changes to explicitly state things to prevent confusion

* Credits page

Created the credits page and added the first photos given

* Extra names

Added the extra names for people that helped contribute to the project

* Updates

Updated docs to provide things that will be useful for competitors

* Finalized suggestions

Made the final suggestions given to update the website

* Changes

Fixed the errors

* Added images

Added extra promo images